+++ src/test/java/domain/game/TestDeckForThreeParticipant.java
@@ -1,8 +1,9 @@
-package domain.deck;
+package domain.game;
 
 import domain.card.Card;
 import domain.card.CardNumber;
 import domain.card.CardShape;
+import domain.deck.DeckStrategy;
 
 import java.util.ArrayList;
 import java.util.List;
